// src/types/partials/MolecularSequenceVariant.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TInt } from '../simpleTypes/Int';
import { TReference } from '../partials/Reference';

export type TMolecularSequenceVariant = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    start?: TInt;
    end?: TInt;
    observedAllele?: String;
    referenceAllele?: String;
    cigar?: String;
    variantPointer?: TReference;
};